Added Compute and Storage Buffer pipelines
- Implemented `PipelineCompute`, `StorageBuffer`, and `StorageBufferSet` abstractions with Vulkan backend support.
- Added methods for creating compute pipelines and storage buffers.
- Updated `Material` to support more uniform types and added array index support for textures.
- Various improvements to shaders, including shadow mapping adjustments and gradient utilities.
- Minor fixes and API enhancements across the engine.

+++ Engine/Source/Renderer/Backend/Vulkan/VulkanImGuiLayer.cpp
@@ -27,6 +27,7 @@ namespace Vanta {
 
 	VulkanImGuiLayer::VulkanImGuiLayer(const std::string& name)
 	{
+
 	}
 
 	VulkanImGuiLayer::~VulkanImGuiLayer()
@@ -35,20 +36,28 @@ namespace Vanta {
 
 	void VulkanImGuiLayer::OnAttach()
 	{
+		// Setup Dear ImGui context
 		IMGUI_CHECKVERSION();
 		ImGui::CreateContext();
 		ImGuiIO& io = ImGui::GetIO(); (void)io;
-		io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard;
-		io.ConfigFlags |= ImGuiConfigFlags_DockingEnable;
-		io.ConfigFlags |= ImGuiConfigFlags_ViewportsEnable;
+		io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard;       // Enable Keyboard Controls
+		//io.ConfigFlags |= ImGuiConfigFlags_NavEnableGamepad;      // Enable Gamepad Controls
+		io.ConfigFlags |= ImGuiConfigFlags_DockingEnable;           // Enable Docking
+		io.ConfigFlags |= ImGuiConfigFlags_ViewportsEnable;         // Enable Multi-Viewport / Platform Windows
+		//io.ConfigViewportsNoDecoration = false;
+		//io.ConfigViewportsNoAutoMerge = true;
+		//io.ConfigViewportsNoTaskBarIcon = true;
 
 		io.Fonts->AddFontFromFileTTF("Resources/Fonts/opensans/OpenSans-Bold.ttf", 18.0f);
 		io.Fonts->AddFontFromFileTTF("Resources/Fonts/opensans/OpenSans-Regular.ttf", 24.0f);
 		io.FontDefault = io.Fonts->AddFontFromFileTTF("Resources/Fonts/opensans/OpenSans-Regular.ttf", 18.0f);
 
+		// Setup Dear ImGui style
 		ImGui::StyleColorsDark();
 		SetDarkThemeColors();
+		//ImGui::StyleColorsClassic();
 
+		// When viewports are enabled we tweak WindowRounding/WindowBg so platform windows can look identical to regular ones.
 		ImGuiStyle& style = ImGui::GetStyle();
 		if (io.ConfigFlags & ImGuiConfigFlags_ViewportsEnable)
 		{
@@ -129,7 +138,6 @@ namespace Vanta {
 			auto device = VulkanContext::GetCurrentDevice()->GetVulkanDevice();
 
 			VK_CHECK_RESULT(vkDeviceWaitIdle(device));
-
 			ImGui_ImplVulkan_Shutdown();
 			ImGui_ImplSDL3_Shutdown();
 			ImGui::DestroyContext();
@@ -217,7 +225,7 @@ namespace Vanta {
 		std::vector<VkCommandBuffer> commandBuffers;
 		commandBuffers.push_back(s_ImGuiCommandBuffers[commandBufferIndex]);
 
-		vkCmdExecuteCommands(drawCommandBuffer, commandBuffers.size(), commandBuffers.data());
+		vkCmdExecuteCommands(drawCommandBuffer, uint32_t(commandBuffers.size()), commandBuffers.data());
 
 		vkCmdEndRenderPass(drawCommandBuffer);
